Refactor MainWindow with Extension
